extract query configuration into separated configuration
